+++ internal/circuitbreaker/manager.go
@@ -16,10 +16,11 @@ package circuitbreaker
 import (
 	"context"
 	"reflect"
+	"sync"
 
 	"github.com/vdaas/vald/internal/errors"
 	"github.com/vdaas/vald/internal/log"
-	"github.com/vdaas/vald/internal/sync"
+	valdsync "github.com/vdaas/vald/internal/sync"
 )
 
 // NOTE: This variable is for observability package.
@@ -36,7 +37,7 @@ type CircuitBreaker interface {
 }
 
 type breakerManager struct {
-	m    sync.Map[string, *breaker]
+	m    valdsync.Map[string, *breaker]
 	opts []BreakerOption
 }