+++ internal/io/io.go
@@ -21,11 +21,11 @@ import (
 	"bytes"
 	"context"
 	"io"
+	"sync"
 
 	"github.com/vdaas/vald/internal/conv"
 	"github.com/vdaas/vald/internal/errors"
 	"github.com/vdaas/vald/internal/safety"
-	"github.com/vdaas/vald/internal/sync"
 )
 
 type (