+++ internal/net/dialer.go
@@ -21,6 +21,7 @@ import (
 	"context"
 	"net"
 	"strconv"
+	"sync"
 	"sync/atomic"
 	"syscall"
 	"time"
@@ -32,7 +33,7 @@ import (
 	"github.com/vdaas/vald/internal/net/control"
 	"github.com/vdaas/vald/internal/observability/trace"
 	"github.com/vdaas/vald/internal/safety"
-	"github.com/vdaas/vald/internal/sync"
+	valdsync "github.com/vdaas/vald/internal/sync"
 	"github.com/vdaas/vald/internal/tls"
 )
 
@@ -59,7 +60,7 @@ type dialer struct {
 	ctrl                  control.SocketController
 	sockFlg               control.SocketFlag
 	dialerDualStack       bool
-	addrs                 sync.Map[string, *addrInfo]
+	addrs                 valdsync.Map[string, *addrInfo]
 	der                   *net.Dialer
 	dialer                func(ctx context.Context, network, addr string) (Conn, error)
 }