+++ pkg/agent/core/ngt/service/option.go
@@ -23,10 +23,10 @@ import (
 	"time"
 
 	core "github.com/vdaas/vald/internal/core/algorithm/ngt"
+	"github.com/vdaas/vald/internal/errgroup"
 	"github.com/vdaas/vald/internal/file"
 	"github.com/vdaas/vald/internal/rand"
 	"github.com/vdaas/vald/internal/strings"
-	"github.com/vdaas/vald/internal/sync/errgroup"
 	"github.com/vdaas/vald/internal/timeutil"
 )